chore: adjust e2e folder
