+++ e2e/cypress/e2e/plugin-clipboard/plugin.cy.ts
@@ -1,7 +1,7 @@
 Cypress.config('baseUrl', `http://localhost:${Cypress.env('SERVER_PORT')}`)
 
 beforeEach(() => {
-  cy.visit('/plugin-clipboard/')
+  cy.visit('/preset-commonmark/')
 })
 
 it('has editor', () => {
