+++ e2e/src/crepe/index.ts
@@ -1,0 +1,4 @@
+export const crepe = {
+  title: 'Crepe',
+  link: '/crepe/',
+}
