+++ e2e/src/multi-editor/index.ts
@@ -1,0 +1,4 @@
+export const multiEditor = {
+  title: 'Multi Editor',
+  link: '/multi-editor/',
+}
