+++ e2e/src/plugin-automd/index.ts
@@ -1,0 +1,4 @@
+export const automd = {
+  title: 'AutoMd',
+  link: '/plugin-automd/',
+}
