--- e2e/src/plugin-clipboard/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Plugin Clipboard</title>
-</head>
-<body>
-  <div id="ui" class="mx-5 my-2 flex justify-between items-center">
-    <div class="flex gap-2">
-      <button id="inspect" class="text-gray-50 rounded px-4 py-2 bg-nord10 hover:bg-nord9">
-        inspect
-      </button>
-      <button id="log" class="text-gray-50 rounded px-4 py-2 bg-nord10 hover:bg-nord9">
-        log
-      </button>
-    </div>
-  </div>
-  <div id="app"></div>
-  <script type="module" src="/plugin-clipboard/main.ts"></script>
-</body>
-</html>
