+++ e2e/src/plugin-listener/index.ts
@@ -1,0 +1,4 @@
+export const listener = {
+  title: 'Listener',
+  link: '/plugin-listener/',
+}
