+++ e2e/src/plugin-math/index.ts
@@ -1,0 +1,4 @@
+export const math = {
+  title: 'Math',
+  link: '/plugin-math/',
+}
