+++ e2e/src/preset-commonmark/index.ts
@@ -1,0 +1,4 @@
+export const commonmark = {
+  title: 'Commonmark',
+  link: '/preset-commonmark/',
+}
