+++ e2e/src/preset-commonmark/main.ts
@@ -1,6 +1,7 @@
 import { Editor, rootCtx } from '@milkdown/core'
 import { nord } from '@milkdown/theme-nord'
 import { commonmark, toggleEmphasisCommand, toggleStrongCommand } from '@milkdown/preset-commonmark'
+import { clipboard } from '@milkdown/plugin-clipboard'
 import { history } from '@milkdown/plugin-history'
 import { callCommand } from '@milkdown/utils'
 
@@ -19,7 +20,9 @@ setup(() => {
     .config(nord)
     .use(commonmark)
     .use(history)
+    .use(clipboard)
     .create()
+
   editor.then((instance) => {
     globalThis.commands.toggleStrong = () => {
       instance.action(callCommand(toggleStrongCommand.key))
