+++ e2e/src/preset-gfm/index.ts
@@ -1,0 +1,4 @@
+export const gfm = {
+  title: 'GFM',
+  link: '/preset-gfm/',
+}
