+++ e2e/stories/components/code-block.stories.ts
@@ -12,7 +12,7 @@ import { keymap } from '@codemirror/view'
 import { html } from 'atomico'
 
 import '@milkdown/theme-nord/style.css'
-import '../style.css'
+import '../../src/style.css'
 import './code-block.css'
 
 const meta: Meta = {
