+++ e2e/stories/components/image-block.css
