+++ e2e/stories/components/inline-image-block.css
