+++ e2e/stories/components/link-tooltip.stories.ts
@@ -15,7 +15,7 @@ import type { EditorView } from '@milkdown/prose/view'
 import type { EditorState } from '@milkdown/prose/state'
 
 import '@milkdown/theme-nord/style.css'
-import '../style.css'
+import '../../src/style.css'
 import './link-tooltip.css'
 
 const meta: Meta = {
