+++ e2e/stories/components/list-item-block.stories.ts
@@ -1,16 +1,17 @@
 import type { Meta, StoryObj } from '@storybook/html'
 import { Editor, defaultValueCtx, editorViewOptionsCtx, rootCtx } from '@milkdown/core'
 import { nord } from '@milkdown/theme-nord'
+import { gfm } from '@milkdown/preset-gfm'
 import { commonmark } from '@milkdown/preset-commonmark'
 import { history } from '@milkdown/plugin-history'
-import { imageBlockComponent } from '@milkdown/components/image-block'
+import { listItemBlockComponent } from '@milkdown/components/list-item-block'
 
 import '@milkdown/theme-nord/style.css'
-import '../style.css'
-import './image-block.css'
+import '../../src/style.css'
+import './list-item-block.css'
 
 const meta: Meta = {
-  title: 'Components/Image Block',
+  title: 'Components/List Item Block',
 }
 
 export default meta
@@ -20,15 +21,31 @@ interface Args {
   defaultValue: string
 }
 
-const logo = `
-![0.5](/milkdown-logo.png)
+const bullet = `
+- List item 1
+    - List item 1.1
+    - List item 1.2
+- List item 2
+- List item 3
 `
 
-const empty = `
-![]()
+const ordered = `
+1. List item 1
+    1. List item 1.1
+    2. List item 1.2
+2. List item 2
+3. List item 3
 `
 
-export const Empty: StoryObj<Args> = {
+const todo = `
+- [ ] Todo list item 1
+    - [ ] Todo List item 1.1
+    - [ ] Todo List item 1.2
+- [ ] Todo list item 2
+- [ ] Todo list item 3
+`
+
+export const Bullet: StoryObj<Args> = {
   render: (args) => {
     const root = document.createElement('div')
     root.classList.add('milkdown-storybook')
@@ -42,7 +59,8 @@ export const Empty: StoryObj<Args> = {
       })
       .config(nord)
       .use(commonmark)
-      .use(imageBlockComponent)
+      .use(gfm)
+      .use(listItemBlockComponent)
       .use(history)
       .create()
 
@@ -50,14 +68,22 @@ export const Empty: StoryObj<Args> = {
   },
   args: {
     readonly: false,
-    defaultValue: empty,
+    defaultValue: bullet,
   },
 }
 
-export const Logo: StoryObj<Args> = {
-  ...Empty,
+export const Ordered: StoryObj<Args> = {
+  ...Bullet,
   args: {
     readonly: false,
-    defaultValue: logo,
+    defaultValue: ordered,
+  },
+}
+
+export const Todo: StoryObj<Args> = {
+  ...Bullet,
+  args: {
+    readonly: false,
+    defaultValue: todo,
   },
 }
