+++ e2e/stories/plugins/listener.stories.ts
@@ -8,7 +8,7 @@ import { expect, userEvent, waitFor, within } from '@storybook/test'
 
 import '@milkdown/theme-nord/style.css'
 
-import '../style.css'
+import '../../src/style.css'
 import './listener.css'
 
 interface Args {
