+++ e2e/vite.config.mts
@@ -1,6 +1,15 @@
 /* Copyright 2021, Milkdown by Mirone. */
 import { resolve } from 'node:path'
 import { defineConfig } from 'vite'
+import { cases } from './src/data'
+
+const data = cases.reduce((acc, { link }) => {
+  const name = link.match(/[\w-]+/)?.[0] as string
+  return {
+    ...acc,
+    [name]: resolve(__dirname, `src${link}index.html`),
+  }
+}, {} as Record<string, string>)
 
 export default defineConfig({
   root: 'src',
@@ -9,15 +18,8 @@ export default defineConfig({
     emptyOutDir: true,
     rollupOptions: {
       input: {
-        'main': resolve(__dirname, 'src/index.html'),
-        'crepe': resolve(__dirname, 'src/crepe/index.html'),
-        'preset-gfm': resolve(__dirname, 'src/preset-gfm/index.html'),
-        'preset-commonmark': resolve(__dirname, 'src/preset-commonmark/index.html'),
-        'plugin-automd': resolve(__dirname, 'src/plugin-automd/index.html'),
-        'plugin-clipboard': resolve(__dirname, 'src/plugin-clipboard/index.html'),
-        'plugin-math': resolve(__dirname, 'src/plugin-math/index.html'),
-        'plugin-listener': resolve(__dirname, 'src/plugin-listener/index.html'),
-        'multi-editor': resolve(__dirname, 'src/multi-editor/index.html'),
+        main: resolve(__dirname, 'src/index.html'),
+        ...data,
       },
     },
   },
